Refactored rebalance annotation and state classes to be part of the api model package
