+++ api/src/main/java/io/strimzi/api/kafka/model/balancing/KafkaRebalanceState.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.api.kafka.operator.assembly;
+package io.strimzi.api.kafka.model.balancing;
 
 public enum KafkaRebalanceState {
     /**
